% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association-dynEqtl.R
\name{dynEqtl}
\alias{dynEqtl}
\title{Gene-variant association for any pair of gene and variant (may be significant or not significant)}
\usage{
dynEqtl(gencodeID = NULL, variantID = NULL, tissueID = NULL,
  datasetID = "gtex_v7")
}
\description{
Calculation based on latest GTEx release
}
