% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association-gene-variant.R
\name{gencodeVariantPair}
\alias{gencodeVariantPair}
\title{Association information}
\usage{
gencodeVariantPair(gencodeID = "ENSG00000105341.14",
  variantID = "rs4674", tissueSiteDetailID = "Adipose_Subcutaneous",
  datasetID = "gtex_v7")
}
\description{
calculates the gene-variant association
}
