% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association-gene-variant.R
\name{gtexAsso.GencodeVariant}
\alias{gtexAsso.GencodeVariant}
\title{Association information}
\source{
https://gtexportal.org/home/api-docs/#/
}
\usage{
gtexAsso.GencodeVariant(gencodeID = "ENSG00000105341.14",
  variantID = "rs4674", tissueSiteDetailID = "Adipose_Subcutaneous",
  datasetID = "gtex_v7")
}
\value{
gene level
check functions 
later
}
\description{
calculates the gene-variant association
}
