% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association-singleTissueEqtl.R
\name{singleTissueEqtl}
\alias{singleTissueEqtl}
\title{Association information per tissue}
\usage{
singleTissueEqtl(snpID = NULL, gencodeID = NULL, geneSymbol = NULL,
  tissueID = NULL, datasetID = NULL)
}
\description{
calculates the gene-variant association
}
